+++ AprgUniqueProblems/AprgUniqueProblems/src/UniqueProblems/BranchAndBound/JobAssignmentProblem.cpp
@@ -9,8 +9,8 @@ namespace alba {
 
 JobAssignmentProblem::JobAssignmentProblem(CostMatrix const& costMatrix)
     : m_costMatrix(costMatrix),
-      m_numberOfWorkers(m_costMatrix.getNumberOfColumns()),
-      m_numberOfJobs(m_costMatrix.getNumberOfRows()),
+      m_numberOfWorkers(static_cast<int>(m_costMatrix.getNumberOfColumns())),
+      m_numberOfJobs(static_cast<int>(m_costMatrix.getNumberOfRows())),
 
       m_nodeId(START_NODE_ID) {}
